+++ Sources/Data/DataContracts/data_sources/MarvelRemoteDataSourceProtocol.swift
@@ -1,5 +1,0 @@
-/// sourcery: AutoMockable
-public protocol MarvelRemoteDataSourceProtocol
-{
-    func fetchHeroes(completion: @escaping @Sendable (CharacterDataContainer) -> Void)
-}