+++ Sources/Data/DataContracts/global_actors/Background.swift
@@ -1,8 +1,0 @@
-public actor BackgroundActor
-{}
-
-@globalActor
-public struct Background
-{
-    public static let shared = BackgroundActor()
-}