+++ Sources/Data/DataContracts/models/CharacterDataModel.swift
@@ -1,7 +1,0 @@
-public struct CharacterDataModel: Codable, Hashable, Identifiable, Sendable
-{
-    public let id: Int
-    public let name: String
-    public let description: String
-    public let thumbnail: Thumbnail
-}