+++ Sources/Domain/CoreDomain/helpers/String+MD5.swift
@@ -1,0 +1,11 @@
+import CryptoKit
+import Foundation
+
+extension String
+{
+    var md5: String
+    {
+        let digest = Insecure.MD5.hash(data: self.data(using: .utf8) ?? Data())
+        return digest.map { .init(format: "%02hhx", $0) }.joined()
+    }
+}