+++ Sources/Domain/CoreDomain/repositories/data_sources/data_models/CharacterDataContainer.swift
@@ -1,0 +1,12 @@
+
+struct CharacterDataContainer: Codable, Hashable, Sendable
+{
+    struct CharacterData: Codable, Hashable, Sendable {
+        let count: Int
+        let limit: Int
+        let offset: Int
+        let results: [CharacterDataModel]
+    }
+
+    let data: CharacterData
+}