+++ Sources/Domain/CoreDomain/repositories/data_sources/data_models/mappers/CharacterModel+CharacterDataModel.swift
@@ -1,0 +1,15 @@
+import CoreDomainContracts
+import Foundation
+
+extension CharacterModel
+{
+    init(_ data: CharacterDataModel)
+    {
+        self.init(
+            id: data.id,
+            name: data.name,
+            heroDescription: data.description,
+            thumbnail: URL(data.thumbnail)
+        )
+    }
+}