+++ Sources/Domain/CoreDomain/repositories/data_sources/data_models/mappers/CharacterModelContainer+CharacterDataContainer.swift
@@ -1,0 +1,14 @@
+import CoreDomainContracts
+
+extension CharacterModelContainer
+{
+    init(_ data: CharacterDataContainer)
+    {
+        self.init(
+            count: data.data.count,
+            limit: data.data.limit,
+            offset: data.data.offset,
+            results: data.data.results.map(CharacterModel.init(_:))
+        )
+    }
+}